requirements.txt
Django==3.2.7
django-allauth==0.45.0
django-phonenumber-field==5.2.0
django-rest-auth==0.9.5
djangorestframework~=3.12.4
django-bootstrap4==3.0.1
phonenumbers==8.12.31
Pillow==8.3.2
psycopg2==2.9.1
